+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## dev
 
+## 8/2/24
+* Adds index_file_path to IGV Sample model (REQUIRES DB MIGRATION)
+
 ## 7/24/24
 * Split RNA Sample models (REQUIRES DB MIGRATION)